Split up shared actions into shared entry actions and shared list actions respectively
