+++ src/sentry/static/sentry/app/redux/actions/workBatch.js
@@ -1,9 +1,8 @@
-import {makeResourceActions} from './shared';
+import {makeResourceActions} from 'app/redux/actions/sharedList';
 
 export const RESOURCE_NAME = 'WORK_BATCH';
 
 export const workBatchActions = makeResourceActions(
   RESOURCE_NAME,
-  '/api/0/organizations/{org}/work-batches/',
-  (org, res) => `/${org.slug}/workbatches/${res.workBatch.id}/`
+  '/api/0/organizations/{org}/work-batches/'
 );
